# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wapul-be",
    "version": "a1e44bde9c77"
  },
  "paths": {
    "/v1/records": {
      "get": {
        "tags": [
          "Records"
        ],
        "summary": "List My Records",
        "operationId": "list_my_records_v1_records_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecordSummary"
                  },
                  "type": "array",
                  "title": "Response List My Records V1 Records Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Records"
        ],
        "summary": "Create Record",
        "description": "기록을 만들고 코드를 블럭으로 나눈 제안과 질문을 함께 돌려준다",
        "operationId": "create_record_v1_records_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/records/{record_id}": {
      "get": {
        "tags": [
          "Records"
        ],
        "summary": "Get Record",
        "operationId": "get_record_v1_records__record_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Records"
        ],
        "summary": "Delete Record",
        "operationId": "delete_record_v1_records__record_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/records/{record_id}/blocks": {
      "put": {
        "tags": [
          "Records"
        ],
        "summary": "Update Blocks",
        "operationId": "update_blocks_v1_records__record_id__blocks_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlocksUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/records/{record_id}/answers": {
      "patch": {
        "tags": [
          "Records"
        ],
        "summary": "Save Answers",
        "description": "답은 사용자가 쓴 그대로 저장한다 (판정·수정 없음)",
        "operationId": "save_answers_v1_records__record_id__answers_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswersUpdate"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/records/{record_id}/layouts": {
      "get": {
        "tags": [
          "Records"
        ],
        "summary": "Get Layouts",
        "operationId": "get_layouts_v1_records__record_id__layouts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LayoutOut"
                  },
                  "title": "Response Get Layouts V1 Records  Record Id  Layouts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/records/{record_id}/groups": {
      "put": {
        "tags": [
          "Records"
        ],
        "summary": "Update Shares",
        "description": "기록을 공유할 그룹을 통째로 정한다. 내가 속한 그룹에만 공유할 수 있다.",
        "operationId": "update_shares_v1_records__record_id__groups_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": "Response Update Shares V1 Records  Record Id  Groups Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerIn": {
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "answer": {
            "type": "string",
            "maxLength": 10000,
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "answer"
        ],
        "title": "AnswerIn"
      },
      "AnswersUpdate": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerIn"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "AnswersUpdate"
      },
      "BlockIn": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/BlockKind"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "start_line": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Start Line"
          },
          "end_line": {
            "type": "integer",
            "minimum": 1.0,
            "title": "End Line"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "name",
          "start_line",
          "end_line"
        ],
        "title": "BlockIn"
      },
      "BlockKind": {
        "type": "string",
        "enum": [
          "input",
          "output",
          "logic"
        ],
        "title": "BlockKind"
      },
      "BlockOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/BlockKind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start_line": {
            "type": "integer",
            "title": "Start Line"
          },
          "end_line": {
            "type": "integer",
            "title": "End Line"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "name",
          "start_line",
          "end_line"
        ],
        "title": "BlockOut"
      },
      "BlocksUpdate": {
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/BlockIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Blocks"
          }
        },
        "type": "object",
        "required": [
          "blocks"
        ],
        "title": "BlocksUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Language": {
        "type": "string",
        "enum": [
          "cpp",
          "python",
          "java"
        ],
        "title": "Language"
      },
      "LayoutOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "markdown": {
            "type": "string",
            "title": "Markdown"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "markdown"
        ],
        "title": "LayoutOut"
      },
...
      "QuestionKind": {
        "type": "string",
        "enum": [
          "problem",
          "logic",
          "boundary",
          "input_meaning",
          "input_condition",
          "output_meaning",
          "output_format",
          "revision",
          "varying"
        ],
        "title": "QuestionKind"
      },
      "QuestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "block_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Id"
          },
          "kind": {
            "$ref": "#/components/schemas/QuestionKind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "examples": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Examples"
          }
        },
        "type": "object",
        "required": [
          "id",
          "block_id",
          "kind",
          "text",
          "answer",
          "examples"
        ],
        "title": "QuestionOut"
      },
      "RecordCreate": {
        "properties": {
          "problem": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Problem"
          },
          "key_idea": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Key Idea"
          },
          "code": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Code"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "initially_wrong": {
            "type": "boolean",
            "title": "Initially Wrong",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "problem",
          "key_idea",
          "code",
          "language"
        ],
        "title": "RecordCreate"
      },
      "RecordOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "key_idea": {
            "type": "string",
            "title": "Key Idea"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "initially_wrong": {
            "type": "boolean",
            "title": "Initially Wrong"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner"
          },
          "group_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Group Ids"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/BlockOut"
            },
            "type": "array",
            "title": "Blocks"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "problem",
          "key_idea",
          "language",
          "owner_name",
          "created_at",
          "updated_at",
          "code",
          "initially_wrong",
          "is_owner",
          "group_ids",
          "blocks",
          "questions"
        ],
        "title": "RecordOut"
      },
      "RecordSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "key_idea": {
            "type": "string",
            "title": "Key Idea"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "problem",
          "key_idea",
          "language",
          "owner_name",
          "created_at",
          "updated_at"
        ],
        "title": "RecordSummary"
      },
...
      "SharesUpdate": {
        "properties": {
          "group_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Group Ids"
          }
        },
        "type": "object",
        "required": [
          "group_ids"
        ],
        "title": "SharesUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}